+++ dropbox-backup/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.0.2 - 2019-01-22
+
+* Show upload progress in 10% increments to limit logging.
+* Option to enable/disable debug logging
+* Set a limit on dropbox use
+
 ## 0.0.1 - 2019-01-19
 
-Initial release
+* Initial release